Updating gopkg.in/urfave/cli.v2 to upstream
