+++ cloud/google/process.go
@@ -35,6 +35,10 @@ func (g *GCPCloud) ProcessList(app string) ([]*pb.Process, error) {
 	return kube.ProcessList(c, ns, app)
 }
 
+func (g *GCPCloud) ProcessSave(name string, strucure map[string]int32) error {
+	return nil
+}
+
 func (g *GCPCloud) latestImage(app *pb.App) string {
 	return fmt.Sprintf("gcr.io/%v/%v:%v", g.Project, app.Name, app.BuildId)
 }
