+++ cloud/google/templates.go
@@ -161,7 +161,7 @@ func AssetNames() []string {
 
 // _bindata is a table, holding each asset generator, mapped to its name.
 var _bindata = map[string]func() (*asset, error){
-	"cloud/google/templates/mysql.tmpl":    cloudGoogleTemplatesMysqlTmpl,
+	"cloud/google/templates/mysql.tmpl": cloudGoogleTemplatesMysqlTmpl,
 	"cloud/google/templates/postgres.tmpl": cloudGoogleTemplatesPostgresTmpl,
 }
 
@@ -204,12 +204,11 @@ type bintree struct {
 	Func     func() (*asset, error)
 	Children map[string]*bintree
 }
-
 var _bintree = &bintree{nil, map[string]*bintree{
 	"cloud": &bintree{nil, map[string]*bintree{
 		"google": &bintree{nil, map[string]*bintree{
 			"templates": &bintree{nil, map[string]*bintree{
-				"mysql.tmpl":    &bintree{cloudGoogleTemplatesMysqlTmpl, map[string]*bintree{}},
+				"mysql.tmpl": &bintree{cloudGoogleTemplatesMysqlTmpl, map[string]*bintree{}},
 				"postgres.tmpl": &bintree{cloudGoogleTemplatesPostgresTmpl, map[string]*bintree{}},
 			}},
 		}},
@@ -262,3 +261,4 @@ func _filePath(dir, name string) string {
 	cannonicalName := strings.Replace(name, "\\", "/", -1)
 	return filepath.Join(append([]string{dir}, strings.Split(cannonicalName, "/")...)...)
 }
+
