+++ cloud/local/deploy.go
@@ -50,6 +50,33 @@ func (g *LocalCloud) ProcessList(app string) ([]*pb.Process, error) {
 	return sched.ProcessList(c, g.Name, app)
 }
 
+func (g *LocalCloud) ProcessSave(name string, structure map[string]int32) error {
+	app, err := g.AppGet(name)
+	if err != nil {
+		return err
+	}
+
+	build, err := g.BuildGet(app.Name, app.BuildId)
+	if err != nil {
+		return err
+	}
+
+	c, err := getKubeClientset(g.Name)
+	if err != nil {
+		return err
+	}
+
+	for key, replicas := range structure {
+		if cmd, ok := build.Procfile[key]; ok {
+			if err := sched.ScalePodReplicas(c, g.Name, getJobID(name, key), g.latestImage(app), cmd, replicas); err != nil {
+				return err
+			}
+		}
+	}
+
+	return nil
+}
+
 func (g *LocalCloud) latestImage(app *pb.App) string {
 	return fmt.Sprintf("%v/%v:%v", g.RegistryAddress, app.Name, app.BuildId)
 }
